Cleaning Walk Through
In this visualization project I wanted to see if there was a correlation between the number of vehicles registered in each state and the Obesity ranking of each state. This hypothesis comes from the thought that if there are more vehicles registered people might be more heavily reliant on those vehicles as their sole means of transportation, which would mean that people commute using healthier means less.

I imported both data sets and used the SELECT * function to get a look at all of the data in both tables. The "LakeCounty.Obesity" table was fairly clean, with the exception of two columns that were not going to be used, so I left it as it was. The "LakeCounty.DMV" table was quite out of sorts, however. The state column had many state names as "% (2)" to represent values. I used the UPDATE function to rename the "% (2)" state names so that I had a more uniform table. I then filtered the data by year=2015, since the data in the "LakeCounty.DMV" table came from 2015. I finally filtered the "LakeCounty.Obesity" table to remove the last two columns, that had no use for this project, and exported them as new .csv files.

This query was done in BigQuery.